LayoutTests/imported/w3c/web-platform-tests/dom/observable/tentative/observable-filter.any-expected.txt

CONSOLE MESSAGE: Error: error while filtering

PASS filter(): Returned Observable filters out results based on predicate
PASS filter(): Errors thrown in filter predicate are emitted to Observer error() handler
PASS filter(): Passes complete() through from source Observable
PASS filter(): Passes error() through from source Observable
PASS filter(): Upon source completion, source Observable teardown sequence happens after downstream filter complete() is called
PASS filter(): Index is passed correctly to predicate
